refactoring: employees gender removed

+++ server/src/application/useCases/products/RemoveProductImagesUseCase/index.ts
@@ -1,4 +1,6 @@
+import { ProductNotFound } from '@app/errors/ProductNotFound'
 import { IProductImagesRepository } from '@app/repositories/IProductImagesRepository'
+import { IProductsRepository } from '@app/repositories/IProductsRepository'
 
 interface Request {
   productId: string
@@ -8,9 +10,18 @@ interface Request {
 type Response = void
 
 export class RemoveProductImagesUseCase {
-  constructor(private productImagesRepo: IProductImagesRepository) {}
+  constructor(
+    private productsRepo: IProductsRepository,
+    private productImagesRepo: IProductImagesRepository,
+  ) {}
 
   async execute({ imagesIds, productId }: Request): Promise<Response> {
+    const productExists = await this.productsRepo.findById(productId)
+
+    if (!productExists) {
+      throw new ProductNotFound()
+    }
+
     await this.productImagesRepo.delete(imagesIds, productId)
   }
 }
